+++ src/providers/ContentShareProvider/state.tsx
@@ -10,7 +10,8 @@ export enum ContentActionType {
   TOGGLE_PAUSE,
   REMOVE,
   DENIED,
-  RESET
+  RESET,
+  UPDATE_MEDIA_URL,
 }
 
 type StartingAction = {
@@ -53,6 +54,11 @@ type ResetAction = {
   payload?: any;
 };
 
+type UpdateMediaUrlAction = {
+  type: ContentActionType.UPDATE_MEDIA_URL;
+  payload?: any;
+};
+
 export type Action =
   | StartingAction
   | DidStopAction
@@ -60,7 +66,8 @@ export type Action =
   | TogglePauseAction
   | RemoveAction
   | DeniedAction
-  | ResetAction;
+  | ResetAction
+  | UpdateMediaUrlAction;
 
 export type ContentShareState = {
   tileId: number | null;
@@ -68,6 +75,7 @@ export type ContentShareState = {
   isLocalShareLoading: boolean;
   isLocalUserSharing: boolean;
   sharingAttendeeId: string | null;
+  mediaUrl: string;
 };
 
 export const initialState: ContentShareState = {
@@ -75,7 +83,8 @@ export const initialState: ContentShareState = {
   paused: false,
   isLocalUserSharing: false,
   isLocalShareLoading: false,
-  sharingAttendeeId: null
+  sharingAttendeeId: null,
+  mediaUrl: '',
 };
 
 export function reducer(
@@ -86,7 +95,7 @@ export function reducer(
     case ContentActionType.STARTING: {
       return {
         ...state,
-        isLocalShareLoading: true
+        isLocalShareLoading: true,
       };
     }
     case ContentActionType.UPDATE: {
@@ -105,7 +114,8 @@ export function reducer(
         tileId: tileState.tileId!,
         isLocalShareLoading: false,
         isLocalUserSharing: isLocalUser,
-        sharingAttendeeId: tileState.boundAttendeeId
+        sharingAttendeeId: tileState.boundAttendeeId,
+        mediaUrl: state.mediaUrl,
       };
     }
     case ContentActionType.REMOVE: {
@@ -128,7 +138,7 @@ export function reducer(
         ...state,
         isLocalShareLoading: false,
         isLocalUserSharing: false,
-        paused: false
+        paused: false,
       };
     }
     case ContentActionType.TOGGLE_PAUSE: {
@@ -138,7 +148,7 @@ export function reducer(
 
       return {
         ...state,
-        paused: !state.paused
+        paused: !state.paused,
       };
     }
     case ContentActionType.DENIED: {
@@ -148,12 +158,18 @@ export function reducer(
 
       return {
         ...state,
-        isLocalShareLoading: false
+        isLocalShareLoading: false,
       };
     }
 
     case ContentActionType.RESET: {
       return initialState;
+    }
+    case ContentActionType.UPDATE_MEDIA_URL: {
+      return {
+        ...state,
+        mediaUrl: payload.mediaUrl,
+      };
     }
     default:
       throw new Error('Incorrect type in VideoProvider');