+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `isSelected` prop to `ControlBarButton` component
 - Add `UpAndDownCaret` icon component
 - Add channel list component for chat
+- Add Share Video File feature
 
 ### Changed